SUMMER PROJECT
This is summer project for Edunova.

MY TASK WAS - and I did everything from the list

1.Use NetBeans to develop application
2. The app must have a menu with buttons / pictures / links to open the
window
3. Set up a link in the menu that opens a new window with the image of
the ERA diagram
4. Have as many windows as you have a table in a base that does not have
any external keys
5. How much you have a table in a base that does not have any external
keys in it
6. Each window must enable CRUD - entry, read, change, and delete all
data that is not tied to some other external key tables.
7. Set to github code
In readme.md write which points are realized
8. Set up a link to open the github code in the menu
9. All code knocks from 0, do not copy / paste
10. Apply package last name.redame

Advanced:
11. To enter and change the database, you need to have control over all
input data
12. Style change the default (default) appearance of the app - do not do
this before solving all previous queues